Trainer skeleton refactored into a separate class
